pyproject.toml
[project]
name = "pokeroast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "maxykoin", email = "stayerim@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "pyttsx3>=2.99",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
]

[build-system]
requires = ["uv_build>=0.9.20,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
